+++ sdk_core/command_handler/command_impl.cpp
@@ -721,7 +721,7 @@ livox_status CommandImpl::SetLivoxLidarPpsSyncMode(uint32_t handle, LivoxLidarPp
 livox_status CommandImpl::SetLivoxLidarEscMode(uint32_t handle, LivoxLidarEscMode esc_mode, LivoxLidarAsyncControlCallback cb, void* client_data) {
   uint8_t req_buff[kMaxCommandBufferSize] = {0};
   uint16_t req_len = 0;
-  
+
   uint16_t key_num = 1;
   memcpy(&req_buff[req_len], &key_num, sizeof(key_num));
   req_len = sizeof(key_num) + sizeof(uint16_t);
@@ -732,6 +732,32 @@ livox_status CommandImpl::SetLivoxLidarEscMode(uint32_t handle, LivoxLidarEscMod
   uint8_t* val_esc_mode = reinterpret_cast<uint8_t*>(&kv->value[0]);
   *val_esc_mode = esc_mode;
   req_len += sizeof(LivoxLidarKeyValueParam) - 1 + sizeof(uint8_t);
+
+  return GeneralCommandHandler::GetInstance().SendCommand(handle,
+                    kCommandIDLidarWorkModeControl,
+                    req_buff,
+                    req_len,
+                    MakeCommandCallback<LivoxLidarAsyncControlResponse>(cb, client_data));
+}
+
+livox_status CommandImpl::SetLivoxLidarImuRange(uint32_t handle, LivoxLidarImuOutRate imu_out_rate,
+    LivoxLidarAccelRange accel_range, LivoxLidarGyroRange gyro_range,
+    LivoxLidarAsyncControlCallback cb, void* client_data) {
+  uint8_t req_buff[kMaxCommandBufferSize] = {0};
+  uint16_t req_len = 0;
+
+  uint16_t key_num = 1;
+  memcpy(&req_buff[req_len], &key_num, sizeof(key_num));
+  req_len = sizeof(key_num) + sizeof(uint16_t);
+
+  LivoxLidarKeyValueParam * kv = (LivoxLidarKeyValueParam *)&req_buff[req_len];
+  kv->key = static_cast<uint16_t>(kKeySetImuRange);
+  kv->length = sizeof(LivoxLidarImuRange);
+  LivoxLidarImuRange* val_imu_range = reinterpret_cast<LivoxLidarImuRange*>(&kv->value[0]);
+  val_imu_range->imu_out_rate = static_cast<uint8_t>(imu_out_rate);
+  val_imu_range->accel_range  = static_cast<uint8_t>(accel_range);
+  val_imu_range->gyro_range   = static_cast<uint8_t>(gyro_range);
+  req_len += sizeof(LivoxLidarKeyValueParam) - 1 + sizeof(LivoxLidarImuRange);
 
   return GeneralCommandHandler::GetInstance().SendCommand(handle,
                     kCommandIDLidarWorkModeControl,